tests/wpt/css-tests/selectors-3_dev/html/css3-selectors-lang-008.htm

<!DOCTYPE html>
<html lang="en"><head>
<meta charset="utf-8">
<title>:lang(en-GB), lang="en-GB"</title>
<link href="mailto:ishida@w3.org" rel="author" title="Richard Ishida">
<link href="http://www.w3.org/TR/css3-selectors/#lang-pseudo" rel="help">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type="text/css">
.test div { width: 50px; }
 		#colonlangcontroltest { color: red; font-weight: bold; width: 400px; }
		#colonlangcontroltest:lang(xx) { display:none; }
#box:lang(en-GB) { width: 100px; }
</style>
</head>
<body>



<div class="test"><div lang="en-GB" id="box">&nbsp;</div></div>
<p lang="xx" id="colonlangcontroltest">This test failed because it relies on :lang for results, but :lang is not supported by this browser.


<!--Notes:
This tests a detail related to :lang support. If :lang is not supported, a message will appear and the test will fail.
-->
<script>
test(function() {
assert_equals(document.getElementById('colonlangcontroltest').offsetWidth, 0)
assert_equals(document.getElementById('box').offsetWidth, 100);
}, "A :lang value with language and region subtags will match a lang attribute value with the same subtags.");
</script>

</p><div id="log"></div>



</body></html>
